dr_flac: Fix a compilation warning.

+++ dr_flac.h
@@ -6445,7 +6445,7 @@ static drflac_bool32 drflac__read_and_decode_metadata(drflac_read_proc onRead, d
                 hasKnownFileSize = DRFLAC_TRUE;
             }
 
-            onSeek(pUserData, runningFilePos, DRFLAC_SEEK_SET);
+            onSeek(pUserData, (int)runningFilePos, DRFLAC_SEEK_SET);    /* Safe cast because runningFilePos should always be 42 at this point. */
         }
     }
